+++ v2ex-iOS/Models/V2TopicModel.m
@@ -124,7 +124,7 @@
                         stringModel.attributedString = subString;
                         
                         NSMutableArray *quotes = [[NSMutableArray alloc] init];
-                        for (int i = lastImageQuoteIndex; i < idx; i ++) {
+                        for (NSInteger i = lastImageQuoteIndex; i < idx; i ++) {
                             SCQuote *otherQuote = self.quoteArray[i];
                             otherQuote.range = (NSMakeRange(otherQuote.range.location - lastStringIndex, otherQuote.range.length));
                             [quotes addObject:self.quoteArray[i]];
@@ -162,7 +162,7 @@
                 stringModel.attributedString = subString;
                 
                 NSMutableArray *quotes = [[NSMutableArray alloc] init];
-                for (int i = lastImageQuoteIndex; i < self.quoteArray.count; i ++) {
+                for (NSInteger i = lastImageQuoteIndex; i < self.quoteArray.count; i ++) {
                     SCQuote *otherQuote = self.quoteArray[i];
                     NSInteger location = otherQuote.range.location - lastStringIndex - stringOffset;
                     if (location >= 0) {
